% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictDiag.R
\name{PredictDiag}
\alias{PredictDiag}
\title{PredictDiag}
\usage{
PredictDiag(WT, WT_ref, diag_ref, Hbvarinats)
}
\arguments{
\item{WT}{sequence of wild-type(HbA beta) protein .fasta file}

\item{WT_ref}{reference list of fragments for wild-type protein (HbA beta)}

\item{diag_ref}{possible diagnostic ions for each AA of Hba beta}

\item{Hbvarinats}{sequences of Hb variants .fasta file}
}
\description{
PredictDiag
}
