% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subdatabese.R
\name{SubDatabase}
\alias{SubDatabase}
\title{SubDatabase}
\usage{
SubDatabase(
  HbDatabase,
  Mshift = -29.97,
  error_Da_start = -0.06,
  error_Da_end = 0.05
)
}
\arguments{
\item{HbDatabase}{list of original database or updated database}

\item{Mshift}{observed mass shift at MS1 level}

\item{error_Da_start}{mass error tolerance in Da}

\item{error_Da_end}{mass error tolerance in Da}
}
\description{
SubDatabase
}
